A lot of changes
- performed an EDA
- build a first model with 1 feature, the accuracy is awfully low and the accuracy changes all the time
I wanted to write something so i can reliably get the most accurate of the select features. looping through every feature, training a model for each one and saving the score to a list.
After i looped through all the features I took the feature that had the highest accuracy and added it to another list then i repeated the first loop a couple of hundred times.
At the end i had a list of the winning feature from each iteration, the feature with the highest occurence is the feature with the highest accuracy.
I doubt that this was productive in any way